Replace Import File location from sidebar into editor
