+++ apps/dotcom/client/public/tla/locales-compiled/en.json
@@ -131,5 +131,6 @@
   "f43fb62093": "Back to tldraw.",
   "fb15c53f22": "Viewer",
   "fd1be3efcf": "Are you sure you want to delete this file?",
-  "fed5ec05eb": "Copied link"
+  "fed5ec05eb": "Copied link",
+  "ffcd73e997": "Import file…"
 }
