+++ apps/dotcom/client/public/tla/locales/en.json
@@ -398,5 +398,8 @@
   },
   "fed5ec05eb": {
     "translation": "Copied link"
+  },
+  "ffcd73e997": {
+    "translation": "Import file…"
   }
 }
